+++ docs/src/pages/ButtonDemoPage.tsx
@@ -15,6 +15,7 @@ import {
 import type { TableColumn } from '@archway/valet';
 import type { ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
+import NavDrawer from '../components/NavDrawer';
 
 /*─────────────────────────────────────────────────────────────*/
 export default function ButtonDemoPage() {
@@ -76,6 +77,7 @@ export default function ButtonDemoPage() {
 
   return (
     <Surface>
+      <NavDrawer />
       <Stack spacing={1} preset="showcaseStack">
         {/* Header ------------------------------------------------------- */}
         <Typography variant="h2" bold>