+++ docs/src/pages/RadioGroupDemo.tsx
@@ -18,6 +18,7 @@ import {
   useTheme,
 } from '@archway/valet';
 import { useNavigate } from 'react-router-dom';
+import NavDrawer from '../components/NavDrawer';
 
 /*─────────────────────────────────────────────────────────────────────────────*/
 /* Style Presets                                                              */
@@ -62,6 +63,7 @@ export default function RadioGroupDemoPage() {
 
   return (
     <Surface>
+      <NavDrawer />
       <Stack
         spacing={1}
         preset="showcaseStack"